+++ tests/Generator/DefaultTimeGeneratorTest.php
@@ -15,6 +15,7 @@ use Ramsey\Uuid\Generator\DefaultTimeGenerator;
 use Ramsey\Uuid\Provider\NodeProviderInterface;
 use Ramsey\Uuid\Provider\TimeProviderInterface;
 use Ramsey\Uuid\Test\TestCase;
+use Ramsey\Uuid\Type\Timestamp;
 
 class DefaultTimeGeneratorTest extends TestCase
 {
@@ -56,11 +57,15 @@ class DefaultTimeGeneratorTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
-        $this->timeProvider = $this->getMockBuilder(TimeProviderInterface::class)->getMock();
         $this->nodeProvider = $this->getMockBuilder(NodeProviderInterface::class)->getMock();
         $this->timeConverter = $this->getMockBuilder(TimeConverterInterface::class)->getMock();
         $this->currentTime = ['sec' => 1458733431, 'usec' => 877449];
         $this->calculatedTime = ['low' => '83cb98e0', 'mid' => '98e0', 'hi' => '03cb'];
+
+        $timestamp = new Timestamp($this->currentTime['sec'], $this->currentTime['usec']);
+        $this->timeProvider = Mockery::mock(TimeProviderInterface::class, [
+            'getTimestamp' => $timestamp,
+        ]);
     }
 
     protected function tearDown(): void
@@ -76,8 +81,6 @@ class DefaultTimeGeneratorTest extends TestCase
         $this->nodeProvider->expects($this->once())
             ->method('getNode')
             ->willReturn('122f80ca9e06');
-        $this->timeProvider->method('currentTime')
-            ->willReturn($this->currentTime);
         $this->timeConverter->expects($this->once())
             ->method('calculateTime')
             ->with($this->currentTime['sec'], $this->currentTime['usec'])
@@ -92,9 +95,6 @@ class DefaultTimeGeneratorTest extends TestCase
 
     public function testGenerateUsesTimeProvidersCurrentTime(): void
     {
-        $this->timeProvider->expects($this->once())
-            ->method('currentTime')
-            ->willReturn($this->currentTime);
         $this->timeConverter->expects($this->once())
             ->method('calculateTime')
             ->with($this->currentTime['sec'], $this->currentTime['usec'])
@@ -109,8 +109,6 @@ class DefaultTimeGeneratorTest extends TestCase
 
     public function testGenerateCalculatesTimeWithConverter(): void
     {
-        $this->timeProvider->method('currentTime')
-            ->willReturn($this->currentTime);
         $this->timeConverter->expects($this->once())
             ->method('calculateTime')
             ->with($this->currentTime['sec'], $this->currentTime['usec'])
@@ -131,8 +129,6 @@ class DefaultTimeGeneratorTest extends TestCase
     {
         $expectedBytes = hex2bin('83cb98e098e003cb8fe2122f80ca9e06');
 
-        $this->timeProvider->method('currentTime')
-            ->willReturn($this->currentTime);
         $this->timeConverter->method('calculateTime')
             ->with($this->currentTime['sec'], $this->currentTime['usec'])
             ->willReturn($this->calculatedTime);
@@ -160,7 +156,6 @@ class DefaultTimeGeneratorTest extends TestCase
      */
     public function testGenerateReturnsBinaryStringInUuidFormat(): void
     {
-        $this->timeProvider->method('currentTime')->willReturn($this->currentTime);
         $this->timeConverter->method('calculateTime')->willReturn($this->calculatedTime);
         $binaryUtils = Mockery::mock('alias:' . BinaryUtils::class);
         $binaryUtils->shouldReceive('applyVersion')->andReturn(971);
@@ -205,8 +200,6 @@ class DefaultTimeGeneratorTest extends TestCase
     public function testGenerateUsesRandomSequenceWhenClockSeqNull(): void
     {
         $randomInt = AspectMock::func('Ramsey\Uuid\Generator', 'random_int', 9622);
-        $this->timeProvider->method('currentTime')
-            ->willReturn($this->currentTime);
         $this->timeConverter->expects($this->once())
             ->method('calculateTime')
             ->with($this->currentTime['sec'], $this->currentTime['usec'])